requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.15
django-rest-framework==0.1.0
djangorestframework==3.15.2
sqlparse==0.5.1
typing-extensions==4.12.2
